backend/requirements.txt
# FlowMind AI - Backend Dependencies

# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Database
pymongo==4.6.0
mongoengine==0.27.0

# AI & Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Time Series Forecasting
prophet==1.1.5

# API Clients & HTTP
requests==2.31.0
google-generativeai==0.3.1

# NLP & Text Analysis
transformers==4.36.0
torch==2.1.0
beautifulsoup4==4.12.2

# Data Processing
python-dateutil==2.8.2

# API Authentication
PyJWT==2.8.0

# Task Queue (for automation)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
sentry-sdk==1.39.1

